WIP - return state exactly as provided in authorization request
This is a work-in-progress attempt to return the state parameter value exactly
as provided to the authorization request.

Currently, state is URL-decoded and later re-encoded, turning
"some+%20encoded%2Fstate" into "some%20%20encoded%2Fstate".

This patch works for GET requests by retrieving state from
request.getQueryString() instad of request.getParameterMap().

It does not yet work for the POST request from the consent page to the auth
endpoint if the state parameter is only placed in the request body.